// src/components/ClaimLedger.module.css.d.ts
declare const styles: {
  claimLedger: string;
  header: string;
  empty: string;
  list: string;
  item: string;
  "status-confirmed": string;
  "status-disputed": string;
  "status-unconfirmed": string;
  idRow: string;
  claimId: string;
  statusSymbol: string;
  textRow: string;
};

export = styles;